refactor: separate-feature-classes
