+++ src/main/java/com/habittracker/api/checkin/repository/CheckInRepository.java
@@ -1,6 +1,8 @@
-package com.habittracker.api.checkin;
+package com.habittracker.api.checkin.repository;
 
 import java.util.UUID;
+
+import com.habittracker.api.checkin.model.CheckInEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CheckInRepository extends JpaRepository<CheckInEntity, UUID> {}
